requirements.txt
datatrove
pyarrow
mosestokenizer
toolwrapper>=0.4.1
loomchild-segment
ctranslate2
sentencepiece
# nvidia-cublas-cu12
# nvidia-cudnn-cu12
